+++ src/KnoqReminder.App/KnoqReminder.App/Services/Reminder/ServiceCollectionExtension.cs
@@ -1,3 +1,4 @@
+using KnoqReminder.App.Configurations;
 using KnoqReminder.Domain.Options;
 using KnoqReminder.Domain.Services.Reminder;
 using KnoqReminder.Utilities.Options;
